+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+2014.05.08
+----------
+* Feature: Flask Admin Router
+* Feature: Include ``url_prefix`` option
+
 2014.05.06
 ----------
 * Feature: Flask extension initialisation